2.1.4.4 - lessc can't handle spaces in paths.
To deal with possible spaces in paths we were using GetShortPathName, but this fails in locked down environs such as AppHarbor as it doesn't have permission to traverse up through parent directories to calculate the short name.

Another thing that didn't work due to AppHarbor permissions was setting the --include-path to '.' and setting the working directory.